+++ R/revdep_preinstall.R
@@ -5,6 +5,13 @@
 #' @param skip If TRUE, packages already in the binary crancache cache will
 #' be skipped.  If FALSE, all packages will be pre-installed.
 #'
+#' @param chunk_size If `1`, packages are installed one by one.
+#' If `2`, packages are installed two by two, and so on.
+#' If `Inf`, all packages are installed in one go.
+#' This affects how often the **crancache** package library cache is
+#' updated. The smaller the chunk size is the more frequently it is
+#' updated.
+#'
 #' @param temp_lib_path (character string) The folder where to
 #' install packages during pre-installation.
 #'
@@ -19,8 +26,9 @@
 #' @importFrom future.apply future_lapply
 #' @importFrom progressr progressor
 #' @importFrom crancache install_packages
+#' @importFrom parallel splitIndices
 #' @export
-revdep_preinstall <- function(pkgs, skip = TRUE, temp_lib_path = revdep_preinstall_libs()[1]) {
+revdep_preinstall <- function(pkgs, skip = TRUE, chunk_size = 16L, temp_lib_path = revdep_preinstall_libs()[1]) {
   oopts <- options(Ncpus = availableCores())
   
   pkgs <- unique(pkgs)
@@ -40,14 +48,19 @@ revdep_preinstall <- function(pkgs, skip = TRUE, temp_lib_path = revdep_preinsta
 
   message("Installing into library: ", sQuote(temp_lib_path))
 
-  ## Install one-by-one to update cache sooner
-  p <- progressor(along = pkgs)
-  void <- lapply(pkgs, FUN = function(pkg) {
-    on.exit(p(pkg))
-    message(sprintf("Pre-installing package %s (Ncpus = %d)", pkg, getOption("Ncpus", 1L)))
-    install_packages(pkg, dependencies = TRUE, lib = temp_lib_path)
-  }) #, future.chunk.size = 1L, future.seed = TRUE)
-  invisible(void)  
+  nbr_of_chunks <- ceiling(length(pkgs) / chunk_size)
+  chunks <- splitIndices(length(pkgs), ncl = nbr_of_chunks)
+  message(sprintf("Installing packages in %d chunks of %d packages each", length(chunks), chunk_size))
+  
+  p <- progressor(along = chunks)
+  for (kk in seq_along(chunks)) {
+    on.exit(p())
+    pkgs_chunk <- pkgs[chunks[[kk]]]
+    message(sprintf("%d/%d. Pre-installing %d packages (%s) (Ncpus = %d)", kk, length(chunks), length(pkgs_chunk), paste(sQuote(pkgs_chunk), collapse = ", "), getOption("Ncpus", 1L)))
+    install_packages(pkgs_chunk, dependencies = TRUE, lib = temp_lib_path)
+  }
+  
+  invisible()  
 }
 
 #' @rdname revdep_preinstall