+++ R/run.R
@@ -68,7 +68,7 @@
 #' Rscript -e revdepcheck.extras::run --args --preinstall-todo
 #'
 #' R_REVDEPCHECK_NUM_WORKERS=1 R_REVDEPCHECK_TIMEOUT=60 Rscript -e revdepcheck.extras::run
-#' TAR_SKIP_CLUSTERMQ=true Rscript -e revdepcheck.extras::run
+#' TAR_SKIP_CLUSTERMQ=true xvfb-run Rscript -e revdepcheck.extras::run
 #' ```
 #' 
 #' @importFrom utils help file_test packageVersion str
@@ -81,7 +81,7 @@ run <- function(pkg = ".", ..., warn = 1L, args = base::commandArgs(trailingOnly
   stopifnot(is.character(pkg), length(pkg) == 1L, !is.na(pkg))
   
   stopifnot(length(warn) == 1L, is.numeric(warn), !is.na(warn), warn >= 0L)
-  oopts <- options(warn = warn)
+  oopts <- options(warn = warn, useFancyQuotes = FALSE)
   on.exit(options(oopts))
 
   if ("--help" %in% args) {